TranslatorPython/pyproject.toml
[project]
name = "linto-translator"
version = "1.0.0"
description = "LinTO Python Translator with anti-flicker pipeline"
requires-python = ">=3.11"
dependencies = [
    "aiomqtt>=2.0",
    "pysbd>=0.3.4",
    "rapidfuzz>=3.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
benchmark = [
    "unbabel-comet>=2.0",
]

[project.scripts]
linto-translator = "translator.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["translator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
